Fix Remember_me feature
Remember me not working because of previous 'fix' of getUsername (replaced `return $this->email` with `return $this->username`), not knowing that TokenBasedRememberMeServices.php use `$user->getUsername()` to create cookie. It creates the cookie with username whereas we use email to log in. Mismatch => Log in fail.

Action : rollback to `return $this->email` in User entity getUsername().